TST_Scripts

+++ MDM/MT_Intune/MT_Intune_scriptgenerator.sh
@@ -8,6 +8,7 @@ fi
 
 csv_file="$1"
 template_file="MT_Intune_example.sh"
+template_TST_file="MT_Intune_example_TST.sh"
 
 # Ellenőrizze, hogy a sablonfájl létezik-e
 if [ ! -f "$template_file" ]; then
@@ -15,27 +16,41 @@ if [ ! -f "$template_file" ]; then
     exit 1
 fi
 
+# Ellenőrizze, hogy a sablonfájl létezik-e
+if [ ! -f "$template_TST_file" ]; then
+    echo "Hiba: A TST sablonfájl ($template_TST_file) nem található."
+    exit 1
+fi
+rm -rf "Scripts"
+rm -rf "TST_Scripts"
+mkdir Scripts
+mkdir TST_Scripts
+
 # Olvassa be a CSV fájlt soronként
 while IFS=',' read -r item_name icon_link rest_of_line; do
     # Trim whitespace from variables
     item_name=$(echo "$item_name" | xargs)
-    icon_link=$(echo "$icon_link" | xargs)
-
+    icon_link="https://raw.githubusercontent.com/TRIMDMSupport/InstallomatorMT/refs/heads/newLabels/MDM/MT_Intune/Icons/${item_name}.png"
     # Hozza létre az új fájl nevét
     new_file="./Scripts/MT_Intune_${item_name}.sh"
+    new_TST_file="./TST_Scripts/MT_Intune_TST_${item_name}.sh"
 
     # Másolja a sablonfájlt az új fájlba
     cp "$template_file" "$new_file"
+    cp "$template_TST_file" "$new_TST_file"
 
     # Cserélje ki az 'item' változó értékét
     # Fontos: A sed parancsban a / jelek problémát okozhatnak, ha az 'item_name' tartalmaz / jelet.
     # Ebben az esetben használjunk másik elválasztó karaktert, pl. #
     sed -i '' "s/^item=\"example\"/item=\"${item_name}\"/" "$new_file"
+    sed -i '' "s/^item=\"example\"/item=\"${item_name}\"/" "$new_TST_file"
 
     # Cserélje ki az 'icon' változó értékét
     sed -i '' "s|^icon=\"link\"|icon=\"${icon_link}\"|" "$new_file"
+    sed -i '' "s|^icon=\"link\"|icon=\"${icon_link}\"|" "$new_TST_file"
 
     echo "Létrehozva és módosítva: $new_file"
+    echo "Létrehozva és módosítva: $new_TST_file"
 
 done < "$csv_file"
